Various UI tweaks and prototype bush Added
Changed the main UI node to a canvas layer to keep the UI separate from the world canvas modifiers, also moved certain elements to a UI element that follows the player.

There is now a bush that signals if the player enters and leaves it. This is reflected in a new UI element.